separate responsibility of statistics into separate object
